finished separating code.
