+++ src/index.js
@@ -1,7 +1,7 @@
 import "./style.scss";
 import { interfaceUI, makeFlop } from "./scripts/showUI";
 import TableDeck from "./scripts/tableDeckClass";
-import { make54 } from "./scripts/card";
+import { make54 } from "./scripts/deckBuilding";
 
 // Debug Commands
 interfaceUI.showUI();
